Move `WordValue` into a separate module.
