+++ cryptol.cabal
@@ -167,6 +167,7 @@ library
                        Cryptol.Backend.SeqMap,
                        Cryptol.Backend.SBV,
                        Cryptol.Backend.What4,
+                       Cryptol.Backend.WordValue,
 
                        Cryptol.Eval,
                        Cryptol.Eval.Concrete,
